==> 00-setup.Rmd <==
---
title: "Setup"
output:
  rmarkdown::html_document:
    toc: true
    toc_float: true
params:
  process: FALSE
---

```{r klippy, echo=FALSE, include=TRUE}
klippy::klippy()
```

<br>

## Software installation

Make sure you have installed:

  - **R** -- [link](https://www.r-project.org/)
  - **R studio** -- [link](https://www.rstudio.com/)

----

## R packages installation

Open R studio and run the following lines to install the packages we are going to use during this tutorial.

```{r, eval=params$process}
install.packages("tidyverse")
install.packages("dada2")
install.packages("phyloseq")
install.packages("DECIPHER")
install.packages("phangorn")
install.packages("ape")
install.packages("seqinr")
```
